Add AlterShardingBindingTableRuleStatementUpdater
